extract `\QIT_CLI_Tests\InjectRequestBuilderMockTrait`
