+++ src/tests/UploadTest.php
@@ -3,7 +3,6 @@
 namespace QIT_CLI_Tests;
 
 use QIT_CLI\App;
-use QIT_CLI\RequestBuilder;
 use QIT_CLI\Upload;
 use QIT_CLI_Tests\ZipBuilderHelper as ZipBuilder;
 use RuntimeException;
@@ -13,6 +12,7 @@ use ZipArchive;
 
 class UploadTest extends QITTestCase {
 	use MatchesSnapshots;
+	use InjectRequestBuilderMockTrait;
 
 	protected $to_delete = [];
 
@@ -35,19 +35,6 @@ class UploadTest extends QITTestCase {
 		}
 
 		$this->to_delete = [];
-	}
-
-	protected function inject_request_builder_mock() {
-		$mock = new class() extends RequestBuilder {
-			public function request(): string {
-				global $_test_post_bodies;
-				$_test_post_bodies[] = $this->post_body;
-
-				return json_encode( [ 'upload_id' => 123 ] );
-			}
-		};
-
-		App::singleton( RequestBuilder::class, $mock );
 	}
 
 	/**
